Sdr fix off by one in interpolation factor counter. update readme for wiki

+++ pipelinec/include/dsp/fir_decim.h
@@ -20,7 +20,7 @@
 PRAGMA_MESSAGE(FUNC_WIRES fir_decim_counter) // Not worried about delay of this func
 uint1_t fir_decim_counter()
 {
-  static uint16_t counter; // TODO size counter?
+  static uint8_t counter; // TODO size counter?
   uint1_t ellapsed; 
   if(counter==(FIR_DECIM_FACTOR-1)){
     counter = 0;
